+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Please file a bug if you notice a violation of semantic versioning.
 ### Added
 - kettle-release: early validation of identical set of copyright years in README.md and CHANGELOG.md, adds current year if missing, aborts on mismatch
 - kettle-release: update KLOC in README.md
+- kettle-release: update Rakefile.example with version and date
 ### Changed
 - kettle-release: print package name and version released as final line
 ### Deprecated